feat(doc): Add the minimum requirements needed for the Exoscale Key in the documentation (#11)
# Description
Add the minimum required permission for the Exoscale Key. 
I think that might be useful to other people who don't want to set
unnecessary permissions.

Please don't hesitate to correct if some of the `operation` are not
needed or some are missing.
Please don't hesitate to modify the policy if you have a way to make it
more secure.

## Checklist
(For exoscale contributors)

* [x] Changelog updated (under *Unreleased* block)
* [ ] Integration testing

## Testing

Tested the permissions with `lego`. They are enough to get a
certificate.

---------

Co-authored-by: Philipp Sauter <46172817+sauterp@users.noreply.github.com>
Co-authored-by: Philipp Sauter <philipp.sauter@exoscale.ch>